refactor: languages and routing - domains aware
