+++ packages/default-theme/src/helpers/pages.js
@@ -1,5 +1,5 @@
 export const PAGE_CHECKOUT = "/checkout"
-export const PAGE_ACCOUNT = "/account/profile"
+export const PAGE_ACCOUNT = "/poland/test/account/profile"
 export const PAGE_LOGIN = "/login"
 export const PAGE_SEARCH = "/search"
 export const PAGE_ORDER_SUCCESS = "/order"
