--- packages/default-theme/src/pages/_lang/account/addresses.vue
@@ -1,9 +1,0 @@
-<template>
-  <nuxt-child />
-</template>
-
-<script>
-export default {
-  name: "UserAddresses",
-}
-</script>
