--- packages/default-theme/src/pages/_lang/account/orders.vue
@@ -1,9 +1,0 @@
-<template>
-  <nuxt-child />
-</template>
-
-<script>
-export default {
-  name: "OrderHistory",
-}
-</script>
