+++ packages/default-theme/src/pages/account/orders/_id.vue
@@ -1,4 +1,28 @@
+<template>
+  <div class="order-view">
+    <SfHeading class="login__header" :level="3" :title="$t('Order details')" />
+    <SwOrderDetails :order-id="orderId" />
+  </div>
+</template>
+
 <script>
-import OrderView from "@/pages/_lang/account/orders/_id"
-export default OrderView
+import { SfHeading } from "@storefront-ui/vue"
+import SwOrderDetails from "@/components/SwOrderDetails"
+
+export default {
+  components: { SfHeading, SwOrderDetails },
+  computed: {
+    orderId() {
+      return this.$router.currentRoute.params.id
+    },
+  },
+}
 </script>
+
+<style lang="scss" scoped>
+.order-view {
+  .sw-order-details {
+    width: 100%;
+  }
+}
+</style>
