+++ packages/nuxt-module/plugins/i18n.js
@@ -3,12 +3,17 @@ import VueI18n from "vue-i18n";
 import Middleware from "./middleware";
 import languagesMap from "sw-plugins/languages";
 import domainsMap from "sw-plugins/domains";
+import { useDomains } from "@shopware-pwa/composables";
 
 const FALLBACK_DOMAIN = "/"; // Replace with configuration
 
 Vue.use(VueI18n);
 
 export default ({ app, store }) => {
+  const { currentDomain, availableDomains } = useDomains(app);
+  // console.warn('availableDomains', availableDomains.value);
+  // console.warn('current domain', currentDomain.value)
+  // console.warn('from i18n', app.domainsRouting)
   // Set i18n instance on app
   // This way we can use it in middleware and pages asyncData/fetch
   app.i18n = new VueI18n({
@@ -44,35 +49,4 @@ export default ({ app, store }) => {
 
     return `${domain}${link}`;
   };
-};
-
-/* Replace this middleware with a domain-centered middleware */
-Middleware.i18n = function ({ isHMR, app, store, route, params, redirect }) {
-  let domainGuess = `/${route.params.lang}`;
-
-  const defaultLocale = app.i18n.fallbackLocale; // Should be renamed to domain
-  // If middleware is called from hot module replacement, ignore it
-  if (isHMR) {
-    return;
-  }
-
-  let domainConfig = domainsMap[domainGuess];
-
-  if (domainConfig === undefined) {
-    domainConfig = domainsMap[FALLBACK_DOMAIN];
-  }
-
-  let locale = domainConfig.languageLocaleCode;
-
-  if (!languagesMap[locale]) {
-    locale = defaultLocale;
-    //   return error({ message: "This page could not be found.", statusCode: 404 });
-  }
-
-  const fromMap = languagesMap[locale];
-  app.$shopwareApiInstance.update({ languageId: fromMap && fromMap.id });
-
-  // Set locale
-  store.commit("SET_LANG", locale);
-  app.i18n.locale = locale;
 };
