+++ packages/nuxt-module/src/locales.ts
@@ -56,16 +56,4 @@ export function extendLocales(
       availableLocales: localeFromFiles,
     },
   });
-
-  /* Remove middleware instantiation here and use the same logic to apply new middleware */
-  /* istanbul ignore next */
-  moduleObject.options.router.middleware =
-    moduleObject.options.router.middleware || [];
-  /* istanbul ignore next */
-  if (typeof moduleObject?.options?.router?.middleware === "string") {
-    moduleObject.options.router.middleware = [
-      moduleObject.options.router.middleware,
-    ];
-  }
-  moduleObject.options.router.middleware.push("i18n");
 }
